+++ src/generateKpoints.f90
@@ -74,8 +74,10 @@ CONTAINS
     endif
     cartShift = matmul(K,kLVshift)
     bicCartShift = cartShift ! The k-grid shift, but bring into cell
-    call bring_into_cell(bicCartShift, K, Kinv, eps)
+    call bring_into_cell(bicCartShift, Kinv, K, eps)
+    
     if (.not. equal(cartShift, bicCartShift, eps)) then
+       
        write(*,*) "WARNING: (generateFullKpointList in generateKpoints.f90)"
        write(*,*) "The given shift was outside the first k-grid cell. By translation"
        write(*,*) "symmetry, this is always equivalent to a shift inside the cell."
@@ -85,6 +87,7 @@ CONTAINS
     S = nint(matmul(Kinv,R))
     ! Find the HNF of S, store it in H (B is the transformation matrix)
     call HermiteNormalForm(S,H,B)
+
     ! This the volume ratio between R and K, i.e., the number of unreduced k-points
     n = determinant(H) 
     a = H(1,1); c = H(2,2); f = H(3,3)
@@ -112,9 +115,9 @@ CONTAINS
        call bring_into_cell(KpList(iKP,:),Rinv,R,eps)
     enddo
 
-!!        do idx = 1,n
-!!       write(*,'("kp (bic): ",3(1x,f8.3))') KpList(idx,:)
-!!    end do
+    ! do idx = 1,n
+    !    write(*,'("kp (bic): ",3(1x,f8.3))') KpList(idx,:)
+    ! end do
 
   END subroutine generateFullKpointList
 
@@ -133,8 +136,9 @@ CONTAINS
   !!<parameter regular="true" name="weights"> "Weights" of k-points (length of each orbit).
   !!</parameter>
   !!<parameter regular="true" name="eps_">Finite precision parameter (optional)</parameter>
-  subroutine symmetryReduceKpointList(K, R, kLVshift, UnreducedKpList, SymOps, ReducedList, &
-       weights, eps_)
+  
+  subroutine symmetryReduceKpointList(K, R, kLVshift, UnreducedKpList, SymOps, &
+       ReducedList, weights, eps_)
     real(dp), intent(in) :: K(3,3), R(3,3) 
     real(dp), intent(in) :: kLVshift(3) 
     real(dp), intent(in) :: UnreducedKpList(:,:) 
@@ -166,13 +170,13 @@ CONTAINS
     call matrix_inverse(K,InvK,err,eps)
     if (err) then
        write(*,*) "ERROR: (symmetryReduceKpointList in generateKpoints.f90)"
-       stop "The k-grid  vectors that were passed in are not linearly dependent."
+       stop "The k-grid vectors that were passed in are linearly dependent."
     END if
 
     call matrix_inverse(R,InvR,err,eps)
     if (err) then
        write(*,*) "ERROR: (symmetryReduceKpointList in generateKpoints.f90)"
-       stop "The reciprocal lattice vectors are not linearly dependent."
+       stop "The reciprocal lattice vectors are linearly dependent."
     END if
 
     !! Check for valid inputs
@@ -194,32 +198,36 @@ CONTAINS
             &reciprocal cell."
        stop
     endif
-
-   shift = matmul(K,kLVshift)
-   ! Integer transformation matrix that takes K to R, not necessarily HNF at the outset
+    
+    ! Put the shift in Cartesian coordinates.
+    shift = matmul(K,kLVshift)
+    ! Integer transformation matrix that takes K to R, not necessarily HNF at the outset
     N = nint(matmul(InvK,R))
     ! Find the HNF of N, store it in H (B is the transformation matrix)
     call HermiteNormalForm(N,H,L)
     ! Left side of transform will be used later, right side (Ri) will not be.
     call SmithNormalForm(H,L,S,Ri)
 
+    ! Diagonal of the SNF.
     D = (/S(1,1),S(2,2),S(3,3)/)
-    
-    cOrbit = 0 ! Count number of unique orbits. I.e., the number of unique k-points
-    nOps = size(SymOps,3) ! Number of symmetry operators in the point group
-    nUR = size(UnreducedKpList,1) ! Number of unreduced k-points
-    hashTable = 0 ! Use a hash table to keep track of symmetrically-equivalent k-points
-    
     call matrix_inverse(K,InvK,err,eps)
+
+    write(*,'(3(1x,i3))') (H(i,:),i=1,3)
+    write(*,*) 
+    write(*,'(3(1x,i3))') (S(i,:),i=1,3)
+    write(*,'("InvK")')
+    write(*,'(3(1x,f7.3))') (InvK(i,:),i=1,3)
+
+    
     if (err) then
        write(*,*) "ERROR: (symmetryReduceKpointList in generateKpoints.f90)"
-       stop "The k-grid  vectors that were passed in are not linearly dependent."
+       stop "The k-grid  vectors that were passed in are linearly dependent."
     END if
 
     call matrix_inverse(R,InvR,err,eps)
     if (err) then
        write(*,*) "ERROR: (symmetryReduceKpointList in generateKpoints.f90)"
-       stop "The reciprocal lattice vectors that were passed in are not linearly dependent."
+       stop "The reciprocal lattice vectors that were passed in are linearly dependent."
     END if
 
     !! Check for valid inputs
@@ -237,20 +245,6 @@ CONTAINS
        write(*,*) "This doesn't make sense. There should be at least 1 kpoint per reciprocal cell."
        stop
     endif
-
-   ! Integer transformation matrix that takes K to R, not necessarily HNF at the outset
-    N = nint(matmul(InvK,R))
-    ! Find the HNF of N, store it in H (B is the transformation matrix)
-    call HermiteNormalForm(N,H,L)
-    ! Left side of transform will be used later, right side (Ri) will not be.
-    call SmithNormalForm(H,L,S,Ri)
-    write(*,'(3(1x,i3))') (H(i,:),i=1,3)
-    write(*,*) 
-    write(*,'(3(1x,i3))') (S(i,:),i=1,3)
-    write(*,'("InvK")')
-    write(*,'(3(1x,f7.3))') (InvK(i,:),i=1,3)
-    
-    D = (/S(1,1),S(2,2),S(3,3)/)
     
     cOrbit = 0 ! Count how many unique orbits there are. I.e., the number of unique kpoints
     nOps = size(SymOps,3) ! Number of symmetry operators in the point group
@@ -260,42 +254,52 @@ CONTAINS
     iWt = 0 ! count the number of members of each orbit
     do iUnRdKpt = 1,nUR ! Loop over each k-point and mark off its symmetry brothers
        urKpt = UnreducedKpList(iUnRdKpt,:) ! unrotated k-point (shorter name for clarity)
-      write(*,'(/,"kpt#: ",i4)') iUnRdKpt
-     idx = findKptIndex(urKpt-shift, InvK, L, D)
-                write(*,'("urKpt: ",3(f6.3,1x),"idx:",i3)') urKpt, idx
-    
-     if (hashTable(idx)/=0) cycle ! This k-point is already on an orbit, skip it
-     cOrbit = cOrbit + 1
-     hashTable(idx) = cOrbit
-     iFirst(cOrbit) = idx
-     iWt(cOrbit) = 1
-     do iOp = 1,nOps ! Loop over each symmetry operator, applying it to each k-point
-        ! Rotate the k-point
-        roKpt = matmul(SymOps(:,:,iOp),urKpt)
-        write(*,'(/,"Operator:",i3)') iOp 
-        write(*,'(3(1x,f7.3))') (SymOps(i,:,iOp),i=1,3)
-        write(*,'("roKpt: ",3(f6.3,1x),i3)') roKpt
-        ! Make sure the rotated k-point is still part of the kgrid. If not, cycle
-        call bring_into_cell(roKpt, InvR, R, eps)
-        write(*,'("icKpt: ",3(f6.3,1x),i3)') roKpt
-        ! Unshift the k-point temporarily to check if it still is a member of the lattice
-        roKpt = roKpt - shift
-        if (.not. equal(matmul(invK,roKpt), nint(matmul(invK,roKpt)), eps)) then
-           cycle
-        endif
-        ! Reshift the k-point before finding its index
-        roKpt = roKpt + shift
-        write(*,'("shKpt: ",3(f6.3,1x),i3)') roKpt
-        idx = findKptIndex(roKpt,InvK,L,D)
-        write(*,'("Op#",1x,i2,5x,"rkpt: ",3(f6.3,1x),5x,"idx: ",i4)') iOp,roKpt,idx
-        ! Is this a new addition to the orbit?
-        if (hashTable(idx)==0) then
-           hashTable(idx) = cOrbit
-           ! if so increase the number of members of this orbit by 1
-           iWt(cOrbit)=iWt(cOrbit)+1 
+       write(*,'(/,"kpt#: ",i4)') iUnRdKpt
+       idx = findKptIndex(urKpt-shift, InvK, L, D)
+       write(*,'("urKpt: ",3(f6.3,1x),"idx:",i3)') urKpt, idx
+       
+       if (hashTable(idx)/=0) cycle ! This k-point is already on an orbit, skip it
+       cOrbit = cOrbit + 1
+       hashTable(idx) = cOrbit
+       ! iFirst(cOrbit) = idx
+       iFirst(cOrbit) = iUnRdKpt
+       iWt(cOrbit) = 1
+       do iOp = 1,nOps ! Loop over each symmetry operator, applying it to each k-point
+          ! Rotate the k-point
+          roKpt = matmul(SymOps(:,:,iOp),urKpt)
+          ! write(*,'(/,"Operator:",i3)') iOp 
+          ! write(*,'(3(1x,f7.3))') (SymOps(i,:,iOp),i=1,3)
+          ! write(*,'("roKpt: ",3(f6.3,1x),i3)') roKpt
+          ! Make sure the rotated k-point is still part of the kgrid. If not, cycle
+          call bring_into_cell(roKpt, InvR, R, eps)
+          ! write(*,'("icKpt: ",3(f6.3,1x),i3)') roKpt
+          ! Unshift the k-point temporarily to check if it still is a member of the lattice
+          roKpt = roKpt - shift
+          if (.not. equal(matmul(invK,roKpt), nint(matmul(invK,roKpt)), eps)) then
+             cycle
           endif
-     enddo
-  enddo                                                  
+          
+          idx = findKptIndex(roKpt, InvK, L, D)
+          ! Reshift the k-point before finding its index
+          roKpt = roKpt + shift
+          ! write(*,'("shKpt: ",3(f6.3,1x),i3)') roKpt
+          ! idx = findKptIndex(roKpt, InvK, L, D)
+          ! write(*,'("Op#",1x,i2,5x,"rkpt: ",3(f6.3,1x),5x,"idx: ",i4)') iOp,roKpt,idx
+          ! Is this a new addition to the orbit?
+          if (hashTable(idx)==0) then
+             write(*,'(/,"Operator:",i3)') iOp 
+             write(*,'(3(1x,f7.3))') (SymOps(i,:,iOp),i=1,3)
+             write(*,'("roKpt: ",3(f6.3,1x),i3)') roKpt
+             write(*,'("Op#",1x,i2,5x,"rkpt: ",3(f6.3,1x),5x,"idx: ",i4)') iOp,roKpt,idx
+             
+             print *,"point added"
+             hashTable(idx) = cOrbit
+             ! if so increase the number of members of this orbit by 1
+             iWt(cOrbit)=iWt(cOrbit)+1
+          endif
+       enddo
+       write(*,'(/,"iWt: ",i4)') iWt(cOrbit)
+    enddo
     write(*,*) "Hash table:"
     do i = 1, nUr
        write(*,'("kpt#:",i3,3x,"index:",i3)') i, hashTable(i)
@@ -307,6 +311,10 @@ CONTAINS
     weights = iWt(1:cOrbit)
     do i = 1, cOrbit
        sum = sum + weights(i)
+
+       write(*,'("corbit: ", i3)') i
+       write(*,'("ifirst: ", i3)') iFirst(i)
+       write(*,'("rep kpt: ",3(f6.3,1x))') UnreducedKpList(iFirst(i),:)
        ReducedList(i,:) = UnreducedKpList(iFirst(i),:)
     enddo
     ! Fail safe checks
@@ -316,6 +324,8 @@ CONTAINS
             &one of the orbits of the symmetry group."
        stop
     endif
+    write(*,'(/,"kpt#: ",i4)') iUnRdKpt
+    
     if (size(UnreducedKpList,1)/real(size(weights)) > size(SymOps,3)) then
        write(*,*) "ERROR: (symmetryReduceKpointList in generateKpoints.f90)"
        write(*,*) "The ratio of unreduced to reduced kpoints is larger than the size &